% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dem_table.R
\name{dem_table}
\alias{dem_table}
\title{Make base demographics table}
\usage{
dem_table(dat, expr)
}
\arguments{
\item{dat}{Data.frame containing classification and demographic variables.}

\item{expr}{Expression of how to populate table. \code{cls ~ .} or
\code{cls ~ var1 + var2}}
}
\value{
Data.frame of demographic table.
}
\description{
Generates demographics table from given data.frame.
}
\details{
\code{dem_table} Generate demographics table.

Generates demographic table from given expression.
}
\examples{

data(iris)
dem_table(iris, Species ~ .)
dem_table(iris, Species ~ Sepal.Width + Sepal.Length)

}
\seealso{
\code{\link[demtables]{make_dem_table}}
}
