% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_dem_table.R
\name{view_dem_table}
\alias{view_dem_table}
\title{View demographics table}
\usage{
view_dem_table(tbl)
}
\arguments{
\item{tbl}{Result from \code{make_dem_table}.}
}
\value{
Displays table in viewer.
}
\description{
Views demographics table from \code{make_dem_table} in RStudio.
}
\details{
\code{view_dem_table} View demographics table from make_dem_table.

Displays table from \code{make_dem_table}
}
\examples{

data(iris)
iristbl <- make_dem_table(iris, Species ~ ., 'Iris')
view_dem_table(iristbl)

}
\seealso{
\code{\link[demtables]{make_dem_table}}
}
